# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for an automatic message sending system.",
        "title": "Automatic Message Sending System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/message-sending": {
            "post": {
                "description": "Starts or stops the automatic message sending process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Start or stop automatic message sending",
                "parameters": [
                    {
                        "enum": [
                            "start",
                            "stop"
                        ],
                        "type": "string",
                        "description": "Action to perform",
                        "name": "action",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid action or already in desired state",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sent-messages": {
            "get": {
                "description": "Retrieves a list of sent messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Retrieve sent messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Message"
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving sent messages",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recipient": {
                    "type": "string"
                },
                "sentAt": {
                    "$ref": "#/definitions/sql.NullTime"
                }
            }
        },
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-ins-auth-key",
            "in": "header"
        }
    }
}